redpanda: split out get_partition_state_handler
