+++ src/v/redpanda/admin/server.cc
@@ -4133,119 +4133,6 @@ void admin_server::register_usage_routes() {
 }
 
 namespace {
-
-void fill_raft_state(
-  ss::httpd::debug_json::partition_replica_state& replica,
-  cluster::partition_state state) {
-    ss::httpd::debug_json::raft_replica_state raft_state;
-    auto& src = state.raft_state;
-    raft_state.node_id = src.node();
-    raft_state.term = src.term();
-    raft_state.offset_translator_state = std::move(src.offset_translator_state);
-    raft_state.group_configuration = std::move(src.group_configuration);
-    raft_state.confirmed_term = src.confirmed_term();
-    raft_state.flushed_offset = src.flushed_offset();
-    raft_state.commit_index = src.commit_index();
-    raft_state.majority_replicated_index = src.majority_replicated_index();
-    raft_state.visibility_upper_bound_index
-      = src.visibility_upper_bound_index();
-    raft_state.last_quorum_replicated_index
-      = src.last_quorum_replicated_index();
-    raft_state.last_snapshot_term = src.last_snapshot_term();
-    raft_state.received_snapshot_bytes = src.received_snapshot_bytes;
-    raft_state.last_snapshot_index = src.last_snapshot_index();
-    raft_state.received_snapshot_index = src.received_snapshot_index();
-    raft_state.has_pending_flushes = src.has_pending_flushes;
-    raft_state.is_leader = src.is_leader;
-    raft_state.is_elected_leader = src.is_elected_leader;
-    if (src.followers) {
-        for (const auto& f : *src.followers) {
-            ss::httpd::debug_json::raft_follower_state follower_state;
-            follower_state.id = f.node();
-            follower_state.last_flushed_log_index = f.last_flushed_log_index();
-            follower_state.last_dirty_log_index = f.last_dirty_log_index();
-            follower_state.match_index = f.match_index();
-            follower_state.next_index = f.next_index();
-            follower_state.expected_log_end_offset
-              = f.expected_log_end_offset();
-            follower_state.heartbeats_failed = f.heartbeats_failed;
-            follower_state.is_learner = f.is_learner;
-            follower_state.ms_since_last_heartbeat = f.ms_since_last_heartbeat;
-            follower_state.last_sent_seq = f.last_sent_seq;
-            follower_state.last_received_seq = f.last_received_seq;
-            follower_state.last_successful_received_seq
-              = f.last_successful_received_seq;
-            follower_state.suppress_heartbeats = f.suppress_heartbeats;
-            follower_state.is_recovering = f.is_recovering;
-            raft_state.followers.push(std::move(follower_state));
-        }
-    }
-    for (const auto& stm : state.raft_state.stms) {
-        ss::httpd::debug_json::stm_state state;
-        state.name = stm.name;
-        state.last_applied_offset = stm.last_applied_offset;
-        state.max_collectible_offset = stm.last_applied_offset;
-        raft_state.stms.push(std::move(state));
-    }
-    if (src.recovery_state) {
-        ss::httpd::debug_json::follower_recovery_state frs;
-        frs.is_active = src.recovery_state->is_active;
-        frs.pending_offset_count = src.recovery_state->pending_offset_count;
-        raft_state.follower_recovery_state = std::move(frs);
-    }
-    replica.raft_state = std::move(raft_state);
-}
-ss::future<result<std::vector<cluster::partition_state>>>
-get_partition_state(model::ntp ntp, cluster::controller& controller) {
-    if (ntp == model::controller_ntp) {
-        return controller.get_controller_partition_state();
-    }
-    return controller.get_topics_frontend().local().get_partition_state(
-      std::move(ntp));
-}
-
-} // namespace
-
-ss::future<ss::json::json_return_type>
-admin_server::get_partition_state_handler(
-  std::unique_ptr<ss::http::request> req) {
-    const model::ntp ntp = parse_ntp_from_request(req->param);
-    auto result = co_await get_partition_state(ntp, *_controller);
-    if (result.has_error()) {
-        throw ss::httpd::server_error_exception(fmt::format(
-          "Error {} processing partition state for ntp: {}",
-          result.error(),
-          ntp));
-    }
-
-    ss::httpd::debug_json::partition_state response;
-    response.ntp = fmt::format("{}", ntp);
-    const auto& states = result.value();
-    for (const auto& state : states) {
-        ss::httpd::debug_json::partition_replica_state replica;
-        replica.start_offset = state.start_offset;
-        replica.committed_offset = state.committed_offset;
-        replica.last_stable_offset = state.last_stable_offset;
-        replica.high_watermark = state.high_water_mark;
-        replica.dirty_offset = state.dirty_offset;
-        replica.latest_configuration_offset = state.latest_configuration_offset;
-        replica.revision_id = state.revision_id;
-        replica.log_size_bytes = state.log_size_bytes;
-        replica.non_log_disk_size_bytes = state.non_log_disk_size_bytes;
-        replica.is_read_replica_mode_enabled
-          = state.is_read_replica_mode_enabled;
-        replica.read_replica_bucket = state.read_replica_bucket;
-        replica.is_remote_fetch_enabled = state.is_remote_fetch_enabled;
-        replica.is_cloud_data_available = state.is_cloud_data_available;
-        replica.start_cloud_offset = state.start_cloud_offset;
-        replica.next_cloud_offset = state.next_cloud_offset;
-        fill_raft_state(replica, std::move(state));
-        response.replicas.push(std::move(replica));
-    }
-    co_return ss::json::json_return_type(std::move(response));
-}
-
-namespace {
 storage::node::disk_type resolve_disk_type(std::string_view name) {
     if (name == "data") {
         return storage::node::disk_type::data;
